+++ dubbo-rpc/dubbo-rpc-rest/src/main/java/org/apache/dubbo/rpc/protocol/rest/filter/context/FilterContext.java
@@ -35,4 +35,8 @@ public interface FilterContext {
     boolean complete();
 
     void setComplete(boolean complete);
+
+     Object getOriginRequest();
+
+    Object getOriginResponse();
 }